requirements.txt
fastapi[standard]
dotenv
pydantic
psycopg2-binary
sqlalchemy>=1.4
pytest-sqlalchemy-mock
